streamlit/requirements.txt
streamlit
azure.storage.blob
pyarrow
pandas